// src/types/topRatedMoviesTypes.ts
export type topRatedMovieType = {
   adult: boolean;
   backdrop_path: string;
   genre_ids: number[];
   id: number;
   original_language: string;
   original_title: string;
   overview: string;
   popularity: number;
   poster_path: string;
   release_date: string;
   title: string;
   video: boolean;
   vote_average: number;
   vote_count: number;
}

export type RootObject = {
   page: number;
   results: topRatedMovieType[];
   total_pages: number;
   total_results: number;
}

export type topRatedMoviesStateType = {
   topRatedMovies: topRatedMovieType[];
   searchQuery: string;
   totalPage: Number;
   currentPage: Number;
}